demo(server): refactor into modules
